Fixed bug where variables weren't being shared with the Mail templates. Fixes #1.
The System\ServicerProvider listens to the same halting event as this plugin (mailer.beforeAddContent), however because it registers itself first and neither listener specifies a priority the listener there is run first. That listener returns a value which triggers the halting action of the event. Simple solution was to bump of the priority of our listener so that it gets run before the System ServiceProvider listener does.